+++ include/opendspx/label.h
@@ -3,7 +3,7 @@
 
 #include <QString>
 
-namespace QDspx {
+namespace opendspx {
 
     struct Label {
         int pos{0};
