+++ include/opendspx/timesignature.h
@@ -1,7 +1,7 @@
 #ifndef OPENDSPX_MODEL_TIMESIGNATURE_H
 #define OPENDSPX_MODEL_TIMESIGNATURE_H
 
-namespace QDspx {
+namespace opendspx {
 
     struct TimeSignature {
         int index{0};
